% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflora_duplicates.R
\name{reflora_duplicates}
\alias{reflora_duplicates}
\title{Flag duplicates within the downloaded REFLORA occurrence data}
\usage{
reflora_duplicates(df = NULL,
                   duplicates = TRUE,
                   remove = FALSE,
                   keep_acronyms = TRUE,
                   verbose = TRUE,
                   save = TRUE,
                   dir = "reflora_duplicates",
                   filename = "reflora_flagged_duplicates")
}
\arguments{
\item{df}{A vector of specific herbarium acronyms (collection code) in
uppercase letters or leave it as \code{NULL} to summarize specimen records
for all REFLORA-hosted herbaria.}

\item{duplicates}{Logical, if \code{FALSE}, a message showing steps when
summarizing specimen records will not be printed in the console in full.}

\item{remove}{Logical, if \code{FALSE}, a message showing steps when
summarizing specimen records will not be printed in the console in full.}

\item{keep_acronyms}{Logical, if \code{FALSE}, a message showing steps when
summarizing specimen records will not be printed in the console in full.}

\item{verbose}{Logical, if \code{FALSE}, a message showing steps when
summarizing specimen records will not be printed in the console in full.}

\item{save}{Logical, if \code{TRUE}, the search results will be saved on disk}

\item{dir}{Pathway to the computer's directory, where the table-formatted
summary will be saved. The default is to create a directory named
 \code{reflora_records}.}

\item{filename}{Name of the output file to be saved. The default is to create
a file entitled \code{reflora_ocurrence_fixed.csv}.}
}
\value{
A dataframe with the information of the chosen taxon from the chosen
REFLORA Herbaria.
}
\description{
Retrieve specific taxon from the Reflora virtual Herbarium at
\href{https://ipt.jbrj.gov.br/reflora}{REFLORA Virtual Herbarium}
hosted by the \href{https://www.gov.br/jbrj}{Rio de Janeiro Botanical Garden}.
}
\examples{
\dontrun{

fam_taxa <- c("Fabaceae", "Ochnaceae")
occur_df <- reflora_records(herbarium = c("ALCB", "HUEFS", "K", "RB"),
                            taxon = fam_taxa,
                            verbose = TRUE,
                            save = FALSE)

reflora_duplicates(df = occur_df,
                   duplicates = TRUE,
                   remove = FALSE,
                   keep_acronyms = TRUE,
                   verbose = TRUE,
                   save = TRUE,
                   dir = "reflora_duplicates",
                   filename = "reflora_flagged_duplicates")
}

}
\seealso{
\code{\link{reflora_records}}
}
\author{
Domingos Cardoso
}
