% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflora_fix.R
\name{reflora_fix}
\alias{reflora_fix}
\title{Fix specific columns from the downloaded REFLORA occurrence data}
\usage{
reflora_fix(df = NULL,
            scientificName = TRUE,
            recordBy = TRUE,
            coordinates = TRUE,
            country = TRUE,
            verbose = TRUE,
            save = TRUE,
            dir = "reflora_fix",
            filename = "reflora_ocurrence_fixed")
}
\arguments{
\item{df}{A vector of specific herbarium acronyms (collection code) in
uppercase letters or leave it as \code{NULL} to summarize specimen records
for all REFLORA-hosted herbaria.}

\item{scientificName}{Logical, if \code{FALSE}, the search will not check for the
most updated version of the REFLORA dwca files. This argument is often used if
you have defined a specific path to previously downloaded REFLORA dwca files
either manually or with function \code{reflora_occurrence}.}

\item{recordBy}{Logical, if \code{FALSE}, the search will not check for the
most updated version of the REFLORA dwca files. This argument is often used if
you have defined a specific path to previously downloaded REFLORA dwca files
either manually or with function \code{reflora_occurrence}.}

\item{coordinates}{Logical, if \code{FALSE}, the search will not check for the
most updated version of the REFLORA dwca files. This argument is often used if
you have defined a specific path to previously downloaded REFLORA dwca files
either manually or with function \code{reflora_occurrence}.}

\item{country}{Logical, if \code{FALSE}, the search will not check for the
most updated version of the REFLORA dwca files. This argument is often used if
you have defined a specific path to previously downloaded REFLORA dwca files
either manually or with function \code{reflora_occurrence}.}

\item{verbose}{Logical, if \code{FALSE}, a message showing steps when
summarizing specimen records will not be printed in the console in full.}

\item{save}{Logical, if \code{TRUE}, the search results will be saved on disk}

\item{dir}{Pathway to the computer's directory, where the table-formatted
summary will be saved. The default is to create a directory named
 \code{reflora_ocurrence}.}

\item{filename}{Name of the output file to be saved. The default is to create
a file entitled \code{reflora_ocurrence_fixed.csv}.}
}
\value{
A dataframe with the information of the chosen taxon from the chosen
REFLORA Herbaria.
}
\description{
Retrieve specific taxon from the Reflora virtual Herbarium at
\href{https://ipt.jbrj.gov.br/reflora}{REFLORA Virtual Herbarium}
hosted by the \href{https://www.gov.br/jbrj}{Rio de Janeiro Botanical Garden}.
}
\examples{
\dontrun{

fam_taxa <- c("Fabaceae", "Ochnaceae")
occur_df <- reflora_occurrence(herbarium = c("ALCB", "HUEFS", "K", "RB"),
                               taxon = fam_taxa,
                               verbose = TRUE,
                               save = FALSE)

reflora_fix(df = occur_df,
            scientificName = TRUE,
            recordBy = TRUE,
            coordinates = TRUE,
            country = TRUE,
            verbose = TRUE,
            save = TRUE,
            dir = "reflora_fix",
            filename = "reflora_ocurrence_fixed")
}

}
\seealso{
\code{\link{reflora_ocurrence}}
}
\author{
Domingos Cardoso & Carlos Calderón
}
