% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflora_occurrence.R
\name{reflora_ocurrence}
\alias{reflora_ocurrence}
\title{Retrieve specific taxon from the downloaded REFLORA collections}
\usage{
reflora_record(path = NULL,
               herbarium = NULL,
               taxon = NULL,
               state = NULL,
               updates = TRUE,
               verbose = TRUE,
               save = TRUE,
               dir = "reflora_ocurrence",
               filename = "reflora_ocurrence_search")
}
\arguments{
\item{herbarium}{A vector of specific herbarium acronyms (collection code) in
uppercase letters or leave it as \code{NULL} to summarize specimen records
for all REFLORA-hosted herbaria.}

\item{taxon}{A vector with the required taxon.}

\item{state}{A vector with the the required Brazilian states}

\item{updates}{Logical, if \code{FALSE}, the search will not check for the
most updated version of the REFLORA dwca files.}

\item{verbose}{Logical, if \code{FALSE}, a message showing steps when
summarizing specimen records will not be printed in the console in full.}

\item{save}{Logical, if \code{TRUE}, the search results will be saved on disk}

\item{dir}{Pathway to the computer's directory, where the table-formatted
summary will be saved. The default is to create a directory named
 \code{reflora_ocurrence}.}

\item{filename}{Name of the output file to be saved. The default is to create
a file entitled \code{reflora_ocurrence_search.csv}.}
}
\value{
A dataframe with the information of the chosen taxon from the chosen
REFLORA Herbaria.
}
\description{
Retrieve specific taxon from the Reflora virtual Herbarium at
\href{https://ipt.jbrj.gov.br/reflora}{REFLORA Virtual Herbarium}
hosted by the \href{https://www.gov.br/jbrj}{Rio de Janeiro Botanical Garden}.
}
\seealso{
\code{\link{reflora_download}}

\code{\link{reflora_parse}}
}
\author{
Carlos Calderón and Domingos Cardoso
}
