% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflora_reorder.R
\name{reflora_reorder}
\alias{reflora_reorder}
\title{Reorder the downloaded REFLORA occurrence data}
\usage{
reflora_reorder(df = NULL,
                reorder = c("taxa", "collector", "area", "year"),
                verbose = TRUE,
                save = TRUE,
                dir = "reflora_reorder",
                filename = "reflora_ocurrence_reordered")
}
\arguments{
\item{df}{A vector of specific herbarium acronyms (collection code) in
uppercase letters or leave it as \code{NULL} to summarize specimen records
for all REFLORA-hosted herbaria.}

\item{reorder}{A vector with order to be considered:
c("taxa", "collector", "area", "year").}

\item{verbose}{Logical, if \code{FALSE}, a message showing steps when
summarizing specimen records will not be printed in the console in full.}

\item{save}{Logical, if \code{TRUE}, the search results will be saved on disk}

\item{dir}{Pathway to the computer's directory, where the table-formatted
summary will be saved. The default is to create a directory named
 \code{reflora_ocurrence}.}

\item{filename}{Name of the output file to be saved. The default is to create
a file entitled \code{reflora_ocurrence_fixed.csv}.}
}
\value{
A dataframe with the information of the chosen taxon from the chosen
REFLORA Herbaria.
}
\description{
Retrieve specific taxon from the Reflora virtual Herbarium at
\href{https://ipt.jbrj.gov.br/reflora}{REFLORA Virtual Herbarium}
hosted by the \href{https://www.gov.br/jbrj}{Rio de Janeiro Botanical Garden}.
}
\examples{
\dontrun{

fam_taxa <- c("Fabaceae", "Ochnaceae")
occur_df <- reflora_occurrence(herbarium = c("ALCB", "HUEFS", "K", "RB"),
                               taxon = fam_taxa,
                               verbose = TRUE,
                               save = FALSE)

reflora_reorder(df = occur_df,
                reorder = c("taxa", "collector", "area", "year"),
                verbose = TRUE,
                save = TRUE,
                dir = "reflora_reorder",
                filename = "reflora_ocurrence_reordered")
}

}
\seealso{
\code{\link{reflora_ocurrence}}
}
\author{
Domingos Cardoso & Carlos Calderón
}
