% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflora_download.R
\name{reflora_download}
\alias{reflora_download}
\title{Download plant specimen records from REFLORA Virtual Herbarium}
\usage{
reflora_download(herbarium = NULL,
                 verbose = TRUE,
                 dir = "reflora_download")
}
\arguments{
\item{herbarium}{A vector of specific herbarium acronyms (collection code) in
uppercase letters or leave it as NULL to download records for all REFLORA herbaria.}

\item{verbose}{Logical, if \code{FALSE}, a message showing steps when downloading
herbarium specimen records will not be printed in the console in full.}

\item{dir}{Pathway to the computer's directory, where the file will be saved.
The default is to create a directory named **reflora_download**
and the search results will be saved within a subfolder named by collection
code.}
}
\value{
Folder with DwC-A files for an specific or all REFLORA-associated herbaria.
}
\description{
Download plant specimen records in Darwin Core Format from any
herbarium at [Herbário Virtual REFLORA](https://ipt.jbrj.gov.br/reflora)
hosted by the [Rio de Janeiro Botanical Garden](https://www.gov.br/jbrj).
}
\examples{
\dontrun{

reflora_download(herbarium = c("ALCB", "HUEFS", "K", "RB", "US"),
                 verbose = TRUE,
                 dir = "downloaded_reflora")
}

}
\seealso{
\code{\link{reflora_summary}}
}
\author{
Domingos Cardoso
}
