% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflora_indets.R
\name{reflora_indets}
\alias{reflora_indets}
\title{Retrieve specific taxon from the downloaded REFLORA collections}
\usage{
reflora_indets(taxon = NULL,
                   herbarium = NULL,
                   state = NULL,
                   path = NULL,
                   updates = TRUE,
                   verbose = TRUE,
                   save = TRUE,
                   dir = "reflora_ocurrence",
                   filename = "reflora_ocurrence_search")
}
\arguments{
\item{taxon}{A vector with the required taxon.}

\item{herbarium}{A vector of specific herbarium acronyms (collection code) in
uppercase letters or leave it as \code{NULL} to summarize specimen records
for all REFLORA-hosted herbaria.}

\item{state}{A vector with the the required Brazilian states.}

\item{level}{A vector with the taxonomic level as \code{FAMILY} or \code{GENUS}
or both.}

\item{path}{Optional; a pathway to the computer's directory, where the REFLORA-downloaded
dwca folders are. If you do not provide a path, the function will download the
most updated version of the REFLORA dwca files.}

\item{updates}{Logical, if \code{FALSE}, the search will not check for the
most updated version of the REFLORA dwca files. This argument is often used if
you have defined a specific path to previously downloaded REFLORA dwca files
either manually or with function \code{reflora_download}.}

\item{verbose}{Logical, if \code{FALSE}, a message showing steps when
summarizing specimen records will not be printed in the console in full.}

\item{save}{Logical, if \code{TRUE}, the search results will be saved on disk}

\item{dir}{Pathway to the computer's directory, where the table-formatted
summary will be saved. The default is to create a directory named
 \code{reflora_ocurrence}.}

\item{filename}{Name of the output file to be saved. The default is to create
a file entitled \code{reflora_ocurrence_search.csv}.}

\item{recordYear}{A vector with the required record year or year range. For example,
\code{"1992"} or \code{c("1992", "2024")}}
}
\value{
A dataframe with the information of the chosen taxon from the chosen
REFLORA Herbaria.
}
\description{
Retrieve specific taxon from the Reflora virtual Herbarium at
\href{https://ipt.jbrj.gov.br/reflora}{REFLORA Virtual Herbarium}
hosted by the \href{https://www.gov.br/jbrj}{Rio de Janeiro Botanical Garden}.
}
\examples{
\dontrun{

fam_taxa <- c("Fabaceae", "Ochnaceae")
reflora_indets(taxon = fam_taxa,
                   verbose = TRUE,
                   save = TRUE,
                   dir = "reflora_ocurrence",
                   filename = "reflora_ocurrence_search")
}

}
\seealso{
\code{\link{reflora_download}}

\code{\link{reflora_parse}}
}
\author{
Carlos Calderón & Domingos Cardoso
}
