% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflora_parse.R
\name{reflora_parse}
\alias{reflora_parse}
\title{Parse Darwin Core Archive files}
\usage{
reflora_parse(path = NULL,
              herbarium = NULL,
              repatriated = TRUE,
              verbose = TRUE)
}
\arguments{
\item{path}{Pathway to the computer's directory, where the REFLORA-downloaded
dwca folders are.}

\item{herbarium}{A vector of specific herbarium acronyms (collection code) in
uppercase letters or leave it as \code{NULL} to parse specimen records for all
herbarim dcwa folders in the defined path directory.}

\item{repatriated}{Logical. If \code{FALSE}, skips downloading records from
REFLORA-associated herbaria that have been repatriated. Default is \code{TRUE}.
Use \code{reflora_summary()} to check which collections are repatriated.
REFLORA aggregates collections from both Brazilian and international herbaria
that hold Brazilian specimens. In this context, “digital repatriation” refers
to making high-resolution images and associated specimen metadata openly
accessible through a Brazilian public infrastructure (HVR/IPT).}

\item{verbose}{Logical, if \code{FALSE}, a message showing steps when
summarizing specimen records will not be printed in the console in full.}
}
\value{
A named list of Darwin Core Archive objects, one per parsed
herbarium collection. Each element corresponds to the output of
\code{finch::dwca_read()} and contains:
\describe{
\item{\code{$data}}{A list of data tables included in the archive,
including a cleaned and standardized \code{occurrence.txt} data frame
with selected Darwin Core fields.}
\item{\code{$meta}}{Metadata extracted from the archive's EML file.}
\item{\code{$files}}{Information about the archive structure and
associated files.}
}

The returned list is named by herbarium collection codes. When available,
associated summary CSV files (generated by \code{reflora_download()})
are appended as additional elements within each herbarium object.
}
\description{
Read Darwin Core Archive (DwC-A) files from any downloaded dwca
folder at \href{https://ipt.jbrj.gov.br/reflora/}{REFLORA Virtual Herbarium}
hosted by the \href{https://www.gov.br/jbrj/pt-br}{Rio de Janeiro Botanical Garden}.
}
\examples{
\dontrun{

reflora_download(herbarium = c("ALCB", "HUEFS", "K", "RB"),
                 verbose = TRUE,
                 dir = "reflora_download")

dwca <- reflora_parse(path = "reflora_download",
                      verbose = TRUE)
}

}
\seealso{
\code{\link{reflora_download}}
}
\author{
Domingos Cardoso
}
