% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflora_parse.R
\name{reflora_parse}
\alias{reflora_parse}
\title{Parse Darwin Core Archive files}
\usage{
reflora_parse(path = NULL,
              verbose = TRUE)
}
\arguments{
\item{path}{Pathway to the computer's directory, where the REFLORA-downloaded
dwca folders are.}

\item{verbose}{Logical, if \code{FALSE}, a message showing steps when
summarizing specimen records will not be printed in the console in full.}
}
\value{
A list of dwca files, including data and associated metada.
}
\description{
Read Darwin Core Archive (DwC-A) files from any downloaded dwca
folder at \href{https://ipt.jbrj.gov.br/reflora}{REFLORA Virtual Herbarium}
hosted by the \href{https://www.gov.br/jbrj}{Rio de Janeiro Botanical Garden}.
}
\examples{
\dontrun{

reflora_download(herbarium = c("ALCB", "HUEFS", "K", "RB"),
                 verbose = TRUE,
                 dir = "reflora_download")

dwca <- reflora_parse(path = "reflora_download",
                      verbose = TRUE)
}

}
\seealso{
\code{\link{reflora_download}}
}
\author{
Carlos Calderón and Domingos Cardoso
}
