% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflora_parse.R
\name{reflora_parse}
\alias{reflora_parse}
\title{Parse Darwin Core Archive files}
\usage{
reflora_parse(path = NULL,
              herbarium = NULL,
              repatriated = TRUE,
              verbose = TRUE)
}
\arguments{
\item{path}{Pathway to the computer's directory, where the REFLORA-downloaded
dwca folders are.}

\item{herbarium}{A vector of specific herbarium acronyms (collection code) in
uppercase letters or leave it as \code{NULL} to parse specimen records for all
herbarim dcwa folders in the defined path directory.}

\item{repatriated}{Logical. If \code{FALSE}, skips downloading records from
REFLORA-associated herbaria that have been repatriated. Default is \code{TRUE}.
Use \code{reflora_summary()} to check which collections are repatriated.}

\item{verbose}{Logical, if \code{FALSE}, a message showing steps when
summarizing specimen records will not be printed in the console in full.}
}
\value{
A list of dwca files, including data and associated metada.
}
\description{
Read Darwin Core Archive (DwC-A) files from any downloaded dwca
folder at \href{https://ipt.jbrj.gov.br/reflora/}{REFLORA Virtual Herbarium}
hosted by the \href{https://www.gov.br/jbrj/pt-br}{Rio de Janeiro Botanical Garden}.
}
\examples{
\dontrun{

reflora_download(herbarium = c("ALCB", "HUEFS", "K", "RB"),
                 verbose = TRUE,
                 dir = "reflora_download")

dwca <- reflora_parse(path = "reflora_download",
                      verbose = TRUE)
}

}
\seealso{
\code{\link{reflora_download}}
}
\author{
Domingos Cardoso
}
