% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reflora_summary.R
\name{reflora_summary}
\alias{reflora_summary}
\title{Summarize current plant specimen records at REFLORA Virtual Herbarium}
\usage{
reflora_summary(herbarium = NULL,
                verbose = TRUE,
                save = TRUE,
                dir = "reflora_summary")
}
\arguments{
\item{herbarium}{A vector of specific herbarium acronyms (collection code) in
uppercase letters or leave it as NULL to summarize specimen records for all
REFLORA-hosted herbaria.}

\item{verbose}{Logical, if \code{FALSE}, a message showing steps when
summarizing specimen records will not be printed in the console in full.}

\item{save}{Logical, if \code{TRUE}, the search results will be saved on disk.}

\item{dir}{Pathway to the computer's directory, where the table-formatted
summary will be saved. The default is to create a directory named
**reflora_summary**.}
}
\value{
A dataframe summarizing current available plant specimen records in
REFLORA Virtual Herbarium.
}
\description{
Summarize current available plant specimen records at
[Herbário Virtual REFLORA](https://ipt.jbrj.gov.br/reflora)
hosted by the [Rio de Janeiro Botanical Garden](https://www.gov.br/jbrj).
}
\examples{
\dontrun{

reflora_summary(herbarium = c("ALCB", "HUEFS", "K", "RB", "US"),
                verbose = TRUE,
                save = TRUE,
                dir = "reflora_summary")
}

}
\seealso{
\code{\link{reflora_download}}
}
\author{
Domingos Cardoso
}
