% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refloraR-package.R
\docType{package}
\name{refloraR-package}
\alias{refloraR}
\alias{refloraR-package}
\title{refloraR: Tools for efficiently retrieving and analyzing plant specimen data from the Herbario Virtual Reflora}
\description{
Provides tools to access and analyze plant specimen records from Reflora Virtual Herbarium, a botanical data repository maintained by the Rio de Janeiro Botanical Garden. Users can retrieve, filter, and summarize specimen data using taxonomic, geographic, or temporal criteria. The package supports downloading structured Darwin Core (DwC-A) archives and integrates with tidyverse workflows for downstream analysis and visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dboslab.github.io/refloraR-website/}
  \item \url{https://github.com/DBOSlab/refloraR}
  \item Report bugs at \url{https://github.com/DBOSlab/refloraR/issues}
}

}
\author{
\strong{Maintainer}: Domingos Cardoso \email{domingoscardoso@jbrj.gov.br} (\href{https://orcid.org/0000-0001-7072-2656}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Domingos Cardoso \email{domingoscardoso@jbrj.gov.br} (\href{https://orcid.org/0000-0001-7072-2656}{ORCID}) [copyright holder]
  \item Carlos Calderón del Cid \email{caldecid@gmail.com} (\href{https://orcid.org/0000-0002-3897-4615}{ORCID}) [copyright holder]
}

}
\keyword{internal}
