% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refloraR-package.R
\docType{package}
\name{refloraR-package}
\alias{refloraR}
\alias{refloraR-package}
\title{refloraR: An R package for exploring plant specimen collections from REFLORA Virtual Herbarium}
\description{
Explore plant specimen collections from REFLORA Virtual Herbarium.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dboslab.github.io/refloraR/}
  \item \url{https://github.com/dboslab/refloraR}
  \item Report bugs at \url{https://github.com/DBOSlab/refloraR/issues}
}

}
\author{
\strong{Maintainer}: Carlos Calderón \email{caldecid@gmail.com} (\href{https://orcid.org/0000-0002-3897-4615}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Domingos Cardoso \email{cardosobot@gmail.com} (\href{https://orcid.org/0000-0001-7072-2656}{ORCID}) [thesis advisor, copyright holder]
}

}
\keyword{internal}
